na's comment voice moves onto the shared chassis, unchanged to the pixel
`DefaultComment` leaves `CommentThread.tsx` for `Comment.tsx`; every consumer
now imports the chassis directly rather than a name that no longer describes it.
Verified by rendering the old component and the new one for the same props,
resolving every `--faction-*-comment-*` through index.css and diffing the style
streams: every difference is a no-op declaration (`inherit` / `0` /
`transparent`) or the foot rule becoming its own zero-height element, which
computes to the same box.
