+++ frontend/src/components/comments/__tests__/mentionPopoverClip.test.tsx
@@ -27,7 +27,7 @@ import { describe, it, expect } from 'vitest'
 import '../../../i18n'
 import type { CharacterOut } from '../../../api/auth'
 import type { CommentProps } from '../shared'
-import { DefaultComment } from '../CommentThread'
+import DefaultComment from '../Comment'
 import CovenComment from '../voices/CovenComment'
 import EphemeristsComment from '../voices/EphemeristsComment'
 import EverymenComment from '../voices/EverymenComment'
